# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/player/login": {
      "post": {
        "description": "Logs a team in with its access code and the shared event PIN; sets the `spm_access_token` cookie.",
        "operationId": "playerLogin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlayerLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerLoginResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Wrong access code or PIN.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBodyDto"
                }
              }
            }
          }
        },
        "summary": "Player login",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/auth/me": {
      "get": {
        "description": "Returns the authenticated team.",
        "operationId": "me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "No valid player cookie.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBodyDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "spm_access_token": []
          }
        ],
        "summary": "Current player",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/admin/login": {
      "post": {
        "description": "Logs an admin in; sets the `spm_admin_token` cookie.",
        "operationId": "adminLogin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminLoginResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Wrong admin credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBodyDto"
                }
              }
            }
          }
        },
        "summary": "Admin login",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/admin/roster/import": {
      "post": {
        "description": "Upserts teams from a list. Codes become the lowercased player access codes.",
        "operationId": "importRoster",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RosterImportBodyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RosterImportResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Case-colliding codes or access-code clashes.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBodyDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "spm_admin_token": []
          }
        ],
        "summary": "Import roster",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/round": {
      "get": {
        "description": "Round status and configuration.",
        "operationId": "getRound",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminRoundStatusResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "spm_admin_token": []
          }
        ],
        "summary": "Round status",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/round/config": {
      "post": {
        "description": "Toggles games and sets wordleAnswer / cardsSeed. Any subset of fields.",
        "operationId": "updateConfig",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminRoundConfigBodyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminRoundStatusResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Round is ACTIVE — config is locked.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBodyDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "spm_admin_token": []
          }
        ],
        "summary": "Update round config",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/round/start": {
      "post": {
        "description": "IDLE or ENDED → ACTIVE. Sets startedAt and expiresAt.",
        "operationId": "startRound",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminRoundStatusResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "spm_admin_token": []
          }
        ],
        "summary": "Start round",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/round/pause": {
      "post": {
        "description": "ACTIVE → PAUSED. Records pausedAt.",
        "operationId": "pauseRound",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminRoundStatusResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "spm_admin_token": []
          }
        ],
        "summary": "Pause round",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/round/resume": {
      "post": {
        "description": "PAUSED → ACTIVE. Shifts expiresAt so pause time is not counted.",
        "operationId": "resumeRound",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminRoundStatusResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "spm_admin_token": []
          }
        ],
        "summary": "Resume round",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/round/end": {
      "post": {
        "description": "ACTIVE or PAUSED → ENDED.",
        "operationId": "endRound",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminRoundStatusResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "spm_admin_token": []
          }
        ],
        "summary": "End round",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/teams": {
      "get": {
        "description": "Roster with per-game session rows and totalScore.",
        "operationId": "listTeams",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminTeamsResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "spm_admin_token": []
          }
        ],
        "summary": "List teams",
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/teams/reset": {
      "post": {
        "description": "Marks finished sessions ABANDONED so the team can replay. In-progress sessions are untouched.",
        "operationId": "resetTeam",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminTeamResetBodyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamResetResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Unknown teamId.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBodyDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "spm_admin_token": []
          }
        ],
        "summary": "Reset team",
        "tags": [
          "admin"
        ]
      }
    }
  },
  "info": {
    "title": "Spider-Man IEEE Event Platform API",
    "description": "REST API for the Spider-Man themed IEEE event platform.\n\nAuth uses HTTP-only cookies: `spm_access_token` (player) and `spm_admin_token` (admin). Player and admin routes are marked with the matching cookie security scheme. See `docs/frontend.md` (integration guide) and `docs/api.md` (full reference).",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api",
      "description": "The platform prefix — all routes live under /api."
    }
  ],
  "components": {
    "securitySchemes": {
      "spm_access_token": {
        "type": "apiKey",
        "in": "cookie",
        "name": "spm_access_token"
      },
      "spm_admin_token": {
        "type": "apiKey",
        "in": "cookie",
        "name": "spm_admin_token"
      }
    },
    "schemas": {
      "PlayerLoginDto": {
        "type": "object",
        "properties": {
          "accessCode": {
            "type": "string",
            "description": "Roster access code (lowercased team code).",
            "example": "teama"
          },
          "pin": {
            "type": "string",
            "description": "Shared event PIN.",
            "example": "1123"
          }
        },
        "required": [
          "accessCode",
          "pin"
        ]
      },
      "TeamSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Team id (cuid).",
            "example": "cmspz6z310001z9fcyi80n24h"
          },
          "code": {
            "type": "string",
            "description": "Roster code, lowercased — doubles as the player access code.",
            "example": "teama"
          },
          "name": {
            "type": "string",
            "description": "Display name.",
            "example": "Team Alpha"
          },
          "memberNames": {
            "description": "Team member names.",
            "example": [
              "A1",
              "A2"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "room": {
            "type": "string",
            "nullable": true,
            "description": "Room label, if any.",
            "example": "A101"
          }
        },
        "required": [
          "id",
          "code",
          "name",
          "memberNames",
          "room"
        ]
      },
      "PlayerLoginResponseDto": {
        "type": "object",
        "properties": {
          "team": {
            "$ref": "#/components/schemas/TeamSummaryDto"
          }
        },
        "required": [
          "team"
        ]
      },
      "ApiErrorBodyDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 409
          },
          "error": {
            "type": "string",
            "example": "Conflict"
          },
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "description": "Human-readable error message(s)."
          }
        },
        "required": [
          "statusCode",
          "error",
          "message"
        ]
      },
...
      "MeResponseDto": {
        "type": "object",
        "properties": {
          "team": {
            "$ref": "#/components/schemas/TeamSummaryDto"
          }
        },
        "required": [
          "team"
        ]
      },
      "AdminLoginDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "Admin username.",
            "example": "admin"
          },
          "password": {
            "type": "string",
            "description": "Admin password.",
            "example": "••••••••"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "AdminLoginResponseDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "example": "admin"
          }
        },
        "required": [
          "role"
        ]
      },
...
      "RosterEntryDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Team code — becomes the lowercased player access code.",
            "example": "TEAMA"
          },
          "name": {
            "type": "string",
            "example": "Team Alpha"
          },
          "memberNames": {
            "description": "1-8 member names.",
            "example": [
              "A1",
              "A2"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "room": {
            "type": "string",
            "nullable": true,
            "example": "A101"
          }
        },
        "required": [
          "code",
          "name",
          "memberNames"
        ]
      },
      "RosterImportBodyDto": {
        "type": "object",
        "properties": {
          "teams": {
            "description": "1-2000 teams.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RosterEntryDto"
            }
          }
        },
        "required": [
          "teams"
        ]
      },
      "RosterImportResponseDto": {
        "type": "object",
        "properties": {
          "created": {
            "type": "number",
            "example": 2
          },
          "updated": {
            "type": "number",
            "example": 0
          },
          "total": {
            "type": "number",
            "example": 2
          }
        },
        "required": [
          "created",
          "updated",
          "total"
        ]
      },
      "AdminRoundStatusResponseDto": {
        "type": "object",
        "properties": {
          "number": {
            "type": "number",
            "example": 1
          },
          "status": {
            "type": "string",
            "enum": [
              "IDLE",
              "ACTIVE",
              "PAUSED",
              "ENDED"
            ]
          },
          "startedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "pausedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "expiresAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "wordleEnabled": {
            "type": "boolean"
          },
          "shadowEnabled": {
            "type": "boolean"
          },
          "cardsEnabled": {
            "type": "boolean"
          },
          "wordleAnswer": {
            "type": "string",
            "nullable": true,
            "description": "Round wordle answer, if set."
          },
          "cardsSeed": {
            "type": "number",
            "nullable": true,
            "description": "Cards board seed, if set."
          }
        },
        "required": [
          "number",
          "status",
          "startedAt",
          "pausedAt",
          "expiresAt",
          "wordleEnabled",
          "shadowEnabled",
          "cardsEnabled",
          "wordleAnswer",
          "cardsSeed"
        ]
      },
      "AdminRoundConfigBodyDto": {
        "type": "object",
        "properties": {
          "wordleEnabled": {
            "type": "boolean"
          },
          "shadowEnabled": {
            "type": "boolean"
          },
          "cardsEnabled": {
            "type": "boolean"
          },
          "wordleAnswer": {
            "type": "string",
            "nullable": true,
            "description": "Exactly 5 letters; null clears.",
            "example": "spide"
          },
          "cardsSeed": {
            "type": "number",
            "nullable": true,
            "description": "Positive integer; null clears.",
            "example": 123
          }
        }
      },
      "AdminSessionRowDto": {
        "type": "object",
        "properties": {
          "game": {
            "type": "string",
            "enum": [
              "WORDLE",
              "SHADOW",
              "CARDS"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "COMPLETED",
              "TIMEOUT",
              "ABANDONED"
            ]
          },
          "startedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "expiresAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "finishedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "score": {
            "type": "number",
            "nullable": true
          },
          "timeMs": {
            "type": "number",
            "nullable": true
          }
        },
        "required": [
          "game",
          "status",
          "startedAt",
          "expiresAt",
          "finishedAt",
          "score",
          "timeMs"
        ]
      },
      "AdminTeamRowDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cmspz6z310001z9fcyi80n24h"
          },
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "memberNames": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "room": {
            "type": "string",
            "nullable": true
          },
          "sessions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminSessionRowDto"
            }
          },
          "totalScore": {
            "type": "number",
            "example": 0
          }
        },
        "required": [
          "id",
          "code",
          "name",
          "memberNames",
          "room",
          "sessions",
          "totalScore"
        ]
      },
      "AdminTeamsResponseDto": {
        "type": "object",
        "properties": {
          "teams": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminTeamRowDto"
            }
          },
          "total": {
            "type": "number",
            "example": 2
          }
        },
        "required": [
          "teams",
          "total"
        ]
      },
      "AdminTeamResetBodyDto": {
        "type": "object",
        "properties": {
          "teamId": {
            "type": "string",
            "description": "Team id (cuid).",
            "example": "cmspz6z310001z9fcyi80n24h"
          },
          "game": {
            "type": "string",
            "enum": [
              "WORDLE",
              "SHADOW",
              "CARDS"
            ],
            "description": "Game to reset; omit to reset all finished games for the team."
          }
        },
        "required": [
          "teamId"
        ]
      },
      "TeamResetResponseDto": {
        "type": "object",
        "properties": {
          "reset": {
            "type": "number",
            "example": 1,
            "description": "Number of sessions reset."
          }
        },
        "required": [
          "reset"
        ]
      }
    }
  }
}